+++ Hero.h
@@ -12,13 +12,13 @@ protected:
 	int experience_per_level;
 	int health_point_bonus_per_level;
 	int damage_bonus_per_level;
-	float cooldown_multiplier_per_level;
+	double cooldown_multiplier_per_level;
 	
 	void levelup();
 
 public:
 
-	Hero(const std::string, int, int, double, int, int, int, float);
+	Hero(const std::string, int, int, double, int, int, int, double);
 
 	int getLevel();
 	int getMaxHealthPoints();
